+++ pytorch_lightning/strategies/dp.py
@@ -21,7 +21,7 @@ from pytorch_lightning.overrides.data_parallel import LightningParallelModule
 from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
 from pytorch_lightning.plugins.precision import PrecisionPlugin
 from pytorch_lightning.strategies.parallel import ParallelStrategy
-from pytorch_lightning.utilities.apply_func import apply_to_collection, move_data_to_device
+from pytorch_lightning.utilities.apply_func import apply_to_collection
 from pytorch_lightning.utilities.enums import _StrategyType
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.types import _METRIC_COLLECTION, STEP_OUTPUT
@@ -80,7 +80,8 @@ class DataParallelStrategy(ParallelStrategy):
             device: The target device
             dataloader_idx: The index of the dataloader to which the batch belongs.
         """
-        return move_data_to_device(batch, device=device or self.root_device)
+        # DataParallel handles the transfer of batch to the device
+        return batch
 
     def _setup_model(self, model: Module) -> DataParallel:
         """Wraps the given model into a :class:`~torch.nn.parallel.DataParallel` module."""